requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==21.2.0
bcc==0.18.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bitarray==2.9.2
blinker==1.7.0
blis==0.7.11
Brlapi==0.8.3
cachetools==5.3.3
catalogue==2.0.10
certifi==2020.6.20
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
cloudpathlib==0.16.0
colbert-ai==0.2.19
colorama==0.4.4
command-not-found==0.3
confection==0.1.4
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
datasets==2.16.1
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.14
dill==0.3.7
dirtyjson==1.0.8
distlib==0.3.8
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.110.1
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2023.10.0
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.41
glfw==2.6.5
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hidpidaemon==18.4.6
hpack==4.0.0
hsluv==5.0.4
httpcore==1.0.3
httplib2==0.20.2
httpx==0.26.0
huggingface-hub==0.20.3
Hypercorn==0.16.0
hyperframe==6.0.1
identify==2.5.33
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==3.2.0
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-experimental==0.0.55
langchain-openai==0.1.0
langchain-text-splitters==0.0.1
langcodes==3.3.0
langsmith==0.1.31
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama-index==0.9.48
llvmlite==0.42.0
louis==3.20.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
nodeenv==1.8.0
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
onnx==1.15.0
openai==1.10.0
orjson==3.9.15
packaging==23.2
pandas==2.2.0
Pillow==9.0.1
platformdirs==4.1.0
pop-transition==1.1.2
preshed==3.0.9
priority==2.0.0
proto-plus==1.23.0
protobuf==4.25.2
py==1.11.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.11.1
pycairo==1.20.1
pycups==2.0.1
pydantic==2.6.1
pydantic_core==2.16.2
pydbus==0.6.0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynndescent==0.5.11
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.7
pyparsing==2.4.7
PyPDF2==3.0.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
Quart==0.19.4
RAGatouille==0.0.7.post2
regex==2023.12.25
repolib==2.2.1
repoman==1.4.0
requests==2.31.0
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.0
rouge==1.0.1
rsa==4.9
ruff==0.1.13
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.12.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentence-transformers==2.5.0
sentencepiece==0.1.99
sessioninstaller==0.0.0
six==1.16.0
skia-python==87.5
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
srsly==2.4.8
starlette==0.37.2
sympy==1.12
systemd-python==234
taskgroup==0.0.0a4
tenacity==8.2.3
texttable==1.6.4
thinc==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.2
transformers==4.37.2
triton==2.2.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.9.0
umap==0.1.1
umap-learn==0.5.5
uritemplate==4.1.1
urllib3==1.26.5
virtualenv==20.25.0
vispy==0.10.0
voyager==2.0.2
wadllib==1.3.6
wasabi==1.1.2
weasel==0.3.4
websocket-client==1.2.3
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
xxhash==3.4.1
yarl==1.9.4
youtube-dl==2021.12.17
youtube-transcript-api==0.6.2
zipp==1.0.0
